+++ internal/store/cache/users.go
@@ -14,7 +14,7 @@ const UserExpTime = time.Hour
 
 type UserCache interface {
 	Set(ctx context.Context, user *store.User) error
-	Get(ctx context.Context, userID int) (*store.User, error)
+	Get(ctx context.Context, userID int64) (*store.User, error)
 }
 type RedisUserCache struct {
 	rdb *redis.Client
@@ -29,7 +29,7 @@ func (r *RedisUserCache) Set(ctx context.Context, user *store.User) error {
 	return r.rdb.Set(ctx, cacheKey, json, UserExpTime).Err()
 }
 
-func (r *RedisUserCache) Get(ctx context.Context, userID int) (*store.User, error) {
+func (r *RedisUserCache) Get(ctx context.Context, userID int64) (*store.User, error) {
 	cacheKey := fmt.Sprintf("user-%d", userID)
 	data, err := r.rdb.Get(ctx, cacheKey).Result()
 	if err == redis.Nil {
